:sparkles: #5 [feat] 코드리뷰 바탕으로 코드 수정
