+++ src/main/java/nutshell/server/controller/AuthController.java
@@ -3,9 +3,9 @@ package nutshell.server.controller;
 import lombok.RequiredArgsConstructor;
 import nutshell.server.annotation.UserId;
 import nutshell.server.dto.auth.JwtTokensDto;
-import nutshell.server.service.token.TokenService;
 import nutshell.server.service.auth.AuthService;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.token.TokenService;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.IOException;
@@ -16,11 +16,10 @@ import java.io.IOException;
 public class AuthController {
 
     private final AuthService authService;
-    private final TokenService tokenService;
 
     @GetMapping("/login/google")
     public ResponseEntity<Void> googleLogin(
-            @RequestParam String code
+            @RequestParam final String code
     ) throws IOException {
         authService.googleLogin(code);
         return ResponseEntity.ok().build();
@@ -28,14 +27,16 @@ public class AuthController {
 
     @PostMapping("/auth/re-issue")
     public ResponseEntity<JwtTokensDto> reissueToken(
-        @RequestHeader("Authorization") String refreshToken
+        @RequestHeader("Authorization") final String refreshToken
     ){
-        return ResponseEntity.ok(tokenService.reissueToken(refreshToken));
+        return ResponseEntity.ok(authService.reissueToken(refreshToken));
     }
 
     @DeleteMapping("/auth/logout")
-    public ResponseEntity<?> logout(@UserId Long userId){ //토큰넣으면 유저아이디 자동으로 찾아와줌
-        tokenService.logout(userId);
+    public ResponseEntity<?> logout(
+            @UserId Long userId
+    ){
+       authService.logout(userId);
         return ResponseEntity.noContent().build();
     }
 }
