+++ src/main/java/nutshell/server/repository/UserRepository.java
@@ -6,6 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findBySerialIdAndEmail(String serialId, String email);
-
+    Optional<User> findBySerialIdAndEmail(final String serialId, final String email);
 }
